# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Router",
    "version": "generated"
  },
  "paths": {
...
    "/api/v1/tools": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "List Tools",
        "operationId": "list_tools_api_v1_tools_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 500
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "name": "operation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "read",
                    "write",
                    "execute",
                    "unknown"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operation"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          },
          {
            "name": "serverId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Serverid"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Enabled"
            }
          },
          {
            "name": "available",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Available"
            }
          },
          {
            "name": "classificationReviewed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Classificationreviewed"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolPageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{tool_id}": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Get Tool",
        "operationId": "get_tool_api_v1_tools__tool_id__get",
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/mcprouter__registry__wire__ToolDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{tool_id}/classification": {
      "patch": {
        "tags": [
          "tools"
        ],
        "summary": "Patch Classification",
        "operationId": "patch_classification_api_v1_tools__tool_id__classification_patch",
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassificationPatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{tool_id}/enable": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Enable Tool",
        "operationId": "enable_tool_api_v1_tools__tool_id__enable_post",
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{tool_id}/disable": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Disable Tool",
        "operationId": "disable_tool_api_v1_tools__tool_id__disable_post",
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/route": {
      "post": {
        "tags": [
          "routing"
        ],
        "summary": "Route",
        "operationId": "route_api_v1_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/tools": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Tools",
        "operationId": "tools_api_v1_analytics_tools_get",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "surfaced",
                "selected",
                "succeeded",
                "failed",
                "selectionRate",
                "successRate",
                "avgRank",
                "exposedTokens",
                "toolName"
              ],
              "type": "string",
              "default": "surfaced",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "tool",
                "skill",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Kind"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 5,
              "pattern": "^[1-9][0-9]{0,3}[hd]$",
              "default": "7d",
              "title": "Window"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolFunnelPageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/tools/{tool_id}": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Tool Detail",
        "operationId": "tool_detail_api_v1_analytics_tools__tool_id__get",
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 42,
              "title": "Tool Id"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 5,
              "pattern": "^[1-9][0-9]{0,3}[hd]$",
              "default": "7d",
              "title": "Window"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/mcprouter__analytics__wire__ToolDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassificationPatchIn": {
        "properties": {
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "categories": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Categories"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "operation": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "read",
                  "write",
                  "execute",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation"
          },
          "requiredScopes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Requiredscopes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ClassificationPatchIn",
        "description": "Human classification override. Omitted fields are left unchanged; an\nempty body approves the current (auto) classification as-is. Either way the\nrecord becomes `classificationReviewed=true`."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RouteBody": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 4000,
            "title": "Query"
          },
          "agentId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentid"
          },
          "maxTools": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxtools"
          },
          "maxServers": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxservers"
          },
          "allowedServers": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 120,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowedservers"
          },
          "maxSkills": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxskills"
          },
          "kinds": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "tool",
                    "skill"
                  ]
                },
                "type": "array",
                "maxItems": 2,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Kinds"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RouteBody"
      },
      "RouteResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/RoutedToolOut"
            },
            "type": "array",
            "title": "Tools"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/RoutedSkillOut"
            },
            "type": "array",
            "title": "Skills"
          },
          "fallback_used": {
            "type": "boolean",
            "title": "Fallback Used"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          },
          "no_match": {
            "type": "boolean",
            "title": "No Match"
          },
          "max_tools_applied": {
            "type": "integer",
            "title": "Max Tools Applied"
          },
          "max_servers_applied": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Servers Applied"
          },
          "max_skills_applied": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Skills Applied"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "tools",
          "skills",
          "fallback_used",
          "latency_ms",
          "no_match",
          "max_tools_applied",
          "max_servers_applied",
          "max_skills_applied",
          "cached"
        ],
        "title": "RouteResponse"
      },
...
      "RoutedToolOut": {
        "properties": {
          "server": {
            "type": "string",
            "title": "Server"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "server",
          "tool",
          "score"
        ],
        "title": "RoutedToolOut"
      },
...
      "ToolFunnelPageOut": {
        "properties": {
          "window": {
            "$ref": "#/components/schemas/WindowOut"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ToolFunnelOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "window",
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ToolFunnelPageOut"
      },
      "ToolOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "serverId": {
            "type": "string",
            "title": "Serverid"
          },
          "serverName": {
            "type": "string",
            "title": "Servername"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "inputSchema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputschema"
          },
          "schemaHash": {
            "type": "string",
            "title": "Schemahash"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "requiredScopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requiredscopes"
          },
          "classificationReviewed": {
            "type": "boolean",
            "title": "Classificationreviewed"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "stats": {
            "$ref": "#/components/schemas/ToolStatsOut"
          },
          "embeddingBackend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embeddingbackend"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "rank": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          }
        },
        "type": "object",
        "required": [
          "id",
          "serverId",
          "serverName",
          "name",
          "description",
          "inputSchema",
          "schemaHash",
          "version",
          "domain",
          "categories",
          "tags",
          "operation",
          "requiredScopes",
          "classificationReviewed",
          "enabled",
          "available",
          "stats",
          "embeddingBackend",
          "createdAt",
          "updatedAt"
        ],
        "title": "ToolOut"
      },
      "ToolPageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ToolOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ToolPageOut"
      },
...
      "mcprouter__analytics__wire__ToolDetailOut": {
        "properties": {
          "window": {
            "$ref": "#/components/schemas/WindowOut"
          },
          "tool": {
            "$ref": "#/components/schemas/ToolFunnelOut"
          },
          "positionCurve": {
            "items": {
              "$ref": "#/components/schemas/RankPointOut"
            },
            "type": "array",
            "title": "Positioncurve"
          },
          "coSurfaced": {
            "items": {
              "$ref": "#/components/schemas/CoSurfacedOut"
            },
            "type": "array",
            "title": "Cosurfaced"
          }
        },
        "type": "object",
        "required": [
          "window",
          "tool",
          "positionCurve",
          "coSurfaced"
        ],
        "title": "ToolDetailOut"
      },
      "mcprouter__registry__wire__ToolDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "serverId": {
            "type": "string",
            "title": "Serverid"
          },
          "serverName": {
            "type": "string",
            "title": "Servername"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "inputSchema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputschema"
          },
          "schemaHash": {
            "type": "string",
            "title": "Schemahash"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "requiredScopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requiredscopes"
          },
          "classificationReviewed": {
            "type": "boolean",
            "title": "Classificationreviewed"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "stats": {
            "$ref": "#/components/schemas/ToolStatsOut"
          },
          "embeddingBackend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embeddingbackend"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "rank": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ToolVersionOut"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "serverId",
          "serverName",
          "name",
          "description",
          "inputSchema",
          "schemaHash",
          "version",
          "domain",
          "categories",
          "tags",
          "operation",
          "requiredScopes",
          "classificationReviewed",
          "enabled",
          "available",
          "stats",
          "embeddingBackend",
          "createdAt",
          "updatedAt",
          "versions"
        ],
        "title": "ToolDetailOut"
      }
    }
  }
}